+++ .eslintrc.js
@@ -1,0 +1,5 @@
+module.exports = {
+  extends: [require.resolve("./src/node.js")],
+
+  root: true,
+}