+++ lint-staged.config.mjs
@@ -1,0 +1,4 @@
+export default {
+  "*.{js,cjs,mjs,ts,tsx}": ["eslint --fix", "prettier --write"],
+  "*.{ts,tsx}": () => "pnpm run type-check",
+}