+++ src/workers.js
@@ -1,0 +1,5 @@
+module.exports = {
+  rules: {
+    "unicorn/prefer-top-level-await": "off",
+  },
+}