+++ CHANGELOG.md
@@ -1,20 +1,194 @@
 # Changelog
 
-# 1.0.0 (2023-03-17)
+# [1.1.0-beta.11](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.10...v1.1.0-beta.11) (2023-03-24)
+
+
+### 🐛 Bug Fixes
+
+* improve playwright config ([bce7d37](https://github.com/ArchiTower/eslint-config-architower/commit/bce7d37))
+
+# [1.1.0-beta.10](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.9...v1.1.0-beta.10) (2023-03-24)
 
 
 ### ✨ Features
 
-* init ([dea310d](https://github.com/ArchiTower/eslint-config-architower/commit/dea310d))
+* use playwright plugin ([917a0fb](https://github.com/ArchiTower/eslint-config-architower/commit/917a0fb))
+
+# [1.1.0-beta.9](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.8...v1.1.0-beta.9) (2023-03-24)
+
+
+### ✨ Features
+
+* another polishing for vitest ([989fc66](https://github.com/ArchiTower/eslint-config-architower/commit/989fc66))
+
+# [1.1.0-beta.8](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.7...v1.1.0-beta.8) (2023-03-24)
+
+
+### ✨ Features
+
+* optimize vitest rules ([3682c0d](https://github.com/ArchiTower/eslint-config-architower/commit/3682c0d))
+
+# [1.1.0-beta.7](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.6...v1.1.0-beta.7) (2023-03-24)
+
+
+### 🐛 Bug Fixes
+
+* wrong passing params in vitest ([b9ba3c1](https://github.com/ArchiTower/eslint-config-architower/commit/b9ba3c1))
+
+# [1.1.0-beta.6](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.5...v1.1.0-beta.6) (2023-03-24)
+
+
+### ✨ Features
+
+* use vitest plugin ([194ba7f](https://github.com/ArchiTower/eslint-config-architower/commit/194ba7f))
+
+
+### 🐛 Bug Fixes
+
+* bug in workers ([9af67d6](https://github.com/ArchiTower/eslint-config-architower/commit/9af67d6))
+
+# [1.1.0-beta.5](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.4...v1.1.0-beta.5) (2023-03-24)
+
+
+### ✨ Features
+
+* new configs for playwright, workers ([7449f8b](https://github.com/ArchiTower/eslint-config-architower/commit/7449f8b))
+* tune up paths & minor rules ([f034961](https://github.com/ArchiTower/eslint-config-architower/commit/f034961))
+
+# [1.1.0-beta.4](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.3...v1.1.0-beta.4) (2023-03-24)
+
+
+### 🐛 Bug Fixes
+
+* etc missing prefix in TS config ([94ef2c9](https://github.com/ArchiTower/eslint-config-architower/commit/94ef2c9))
+
+# [1.1.0-beta.3](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.2...v1.1.0-beta.3) (2023-03-24)
+
+
+### ✨ Features
+
+* more progress on vitest ([732f717](https://github.com/ArchiTower/eslint-config-architower/commit/732f717))
+
+# [1.1.0-beta.2](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.1...v1.1.0-beta.2) (2023-03-24)
+
+
+### ✨ Features
+
+* split config into smaller parts ([68ede66](https://github.com/ArchiTower/eslint-config-architower/commit/68ede66))
+* wip on vitest config ([14f4f60](https://github.com/ArchiTower/eslint-config-architower/commit/14f4f60))
 
 
 ### 🎫 Chores
 
-* husky & release config ([5061ad3](https://github.com/ArchiTower/eslint-config-architower/commit/5061ad3))
+* **release**: v1.1.0-beta.2 [skip ci] ([c2eedfa](https://github.com/ArchiTower/eslint-config-architower/commit/c2eedfa))
+* upgrade deps ([8339580](https://github.com/ArchiTower/eslint-config-architower/commit/8339580))
+
+
+### 🐛 Bug Fixes
+
+* problem with TSConfig ([71b9d5b](https://github.com/ArchiTower/eslint-config-architower/commit/71b9d5b))
+
+
+### 📝 Documentation
+
+* update README.md ([4171714](https://github.com/ArchiTower/eslint-config-architower/commit/4171714))
 
 
 ### 🔧 Continuous Integration
 
-* create ci config ([a32d343](https://github.com/ArchiTower/eslint-config-architower/commit/a32d343))
+* add check with ESLint before release ([ac0705a](https://github.com/ArchiTower/eslint-config-architower/commit/ac0705a))
+
+# [1.1.0-beta.2](https://github.com/ArchiTower/eslint-config-architower/compare/v1.1.0-beta.1...v1.1.0-beta.2) (2023-03-24)
+
+
+### ✨ Features
+
+* split config into smaller parts ([68ede66](https://github.com/ArchiTower/eslint-config-architower/commit/68ede66))
+
+
+### 🎫 Chores
+
+* upgrade deps ([a8df782](https://github.com/ArchiTower/eslint-config-architower/commit/a8df782))
+
+
+### 🐛 Bug Fixes
+
+* problem with TSConfig ([b060b14](https://github.com/ArchiTower/eslint-config-architower/commit/b060b14))
+
+
+### 📝 Documentation
+
+* update README.md ([4171714](https://github.com/ArchiTower/eslint-config-architower/commit/4171714))
+
+
+### 🔧 Continuous Integration
+
+* add check with ESLint before release ([c2d35bc](https://github.com/ArchiTower/eslint-config-architower/commit/c2d35bc))
+
+# [1.1.0-beta.1](https://github.com/ArchiTower/eslint-config-architower/compare/v1.0.0...v1.1.0-beta.1) (2023-03-23)
+
+
+### ♻ Code Refactoring
+
+* move releaserc to release.config ([db9eb0b](https://github.com/ArchiTower/eslint-config-architower/commit/db9eb0b))
+
+
+### ✅ Tests
+
+* create local ESLint config ([bdc1968](https://github.com/ArchiTower/eslint-config-architower/commit/bdc1968))
+
+
+### ✨ Features
+
+* create config for TS, Node, YAML, JSON & ES ([f38a073](https://github.com/ArchiTower/eslint-config-architower/commit/f38a073))
+* optimize ESLint rules ([25b0ee8](https://github.com/ArchiTower/eslint-config-architower/commit/25b0ee8))
+
+
+### 🎫 Chores
+
+* add dependencies ([67c6d0e](https://github.com/ArchiTower/eslint-config-architower/commit/67c6d0e))
+* add lint-staged ([e86b914](https://github.com/ArchiTower/eslint-config-architower/commit/e86b914))
+* add recommended VSCode extensions ([fc689ad](https://github.com/ArchiTower/eslint-config-architower/commit/fc689ad))
+
+
+### 💄 Styles
+
+* fix a internal problems ([a4c3ba3](https://github.com/ArchiTower/eslint-config-architower/commit/a4c3ba3))
+* format with prettier ([fd89578](https://github.com/ArchiTower/eslint-config-architower/commit/fd89578))
+
+
+### 📝 Documentation
+
+* add minor updates in package.json ([06a547f](https://github.com/ArchiTower/eslint-config-architower/commit/06a547f))
+* add README & LICENSE ([8dc8c0d](https://github.com/ArchiTower/eslint-config-architower/commit/8dc8c0d))
+* update README.md ([eb01257](https://github.com/ArchiTower/eslint-config-architower/commit/eb01257))
+
+
+### 🔧 Continuous Integration
+
+* add dependabot ([a3704ca](https://github.com/ArchiTower/eslint-config-architower/commit/a3704ca))
+* add renovate ([9718057](https://github.com/ArchiTower/eslint-config-architower/commit/9718057))
+* add scanning for GHA ([72af19f](https://github.com/ArchiTower/eslint-config-architower/commit/72af19f))
+* clean up mess ([7bb8f37](https://github.com/ArchiTower/eslint-config-architower/commit/7bb8f37))
+* pnpm cache ([6a0d84c](https://github.com/ArchiTower/eslint-config-architower/commit/6a0d84c))
+* replace dependabot with renovate ([cb1b406](https://github.com/ArchiTower/eslint-config-architower/commit/cb1b406))
+* set beta releases from develop ([9d74c39](https://github.com/ArchiTower/eslint-config-architower/commit/9d74c39))
+
+# 1.0.0 (2023-03-17)
+
+### ✨ Features
+
+- init
+  ([dea310d](https://github.com/ArchiTower/eslint-config-architower/commit/dea310d))
+
+### 🎫 Chores
+
+- husky & release config
+  ([5061ad3](https://github.com/ArchiTower/eslint-config-architower/commit/5061ad3))
+
+### 🔧 Continuous Integration
+
+- create ci config
+  ([a32d343](https://github.com/ArchiTower/eslint-config-architower/commit/a32d343))
 
 ## 1.0.1 (2023-03-14)